requirements.txt
requests
shapely
python-dotenv